% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PMO.R
\name{PMO}
\alias{GetOHLCSqlite}
\alias{KalmanFilter}
\alias{Kama}
\alias{MesaStochastic}
\alias{PMO}
\alias{SuperSmoother}
\alias{TTREx}
\title{TTREx}
\usage{
PMO(x, lookBackLen = 35, smoothLen = 20)

MesaStochastic(x, n = 20)

SuperSmoother(x, emaLength = 5)

GetOHLCSqlite(fn, ticker, startSecs, endSecs, numMinutes = 1)

KalmanFilter(S, gain = 500)

Kama(s, fast = 2, slow = 30, period = 10, debug = FALSE)
}
\arguments{
\item{x}{a xts data object of prices}

\item{lookBackLen}{length of lookback period to be used in the first EMA.}

\item{smoothLen}{number of periods to smooth the result of the first EMA.}
}
\value{
For \code{PMO()} a xts object that contains the following columns:
\enumerate{
 \item \code{roc1} The discrete day rate-of-change
 \item \code{ema35} The \code{ema(35, roc1) * 10}
 \item \code{ema20} The \code{ema(20, ema35)}
}
For \code{MesaStochastic()} a xts object that contains the following columns:
\enumerate{
 \item \code{HP}  The High Pass Filter
 \item \code{SuperSmoother} Ehler's Super Smoother on the High Pass Filter
 \item \code{Stoch}         Regular Stochastics
 \item \code{MesaStoch}     Ehler's Stochastics
}
For \code{SuperSmoother()} a xts object that contains the following columns:
\enumerate{
 \item \code{SS}      Ehler's SuperSmoother
 \item \code{SSEMA}   An EMA of the SuperSmoother
 \item \code{SSDiff}  The percent difference of SS and SSEMA (\code{(SS-SSEMA)/SS)}
 \item \code{SSSig}   An EMA of SSDiff
}
For \code{GetOHLCSqlite()} Load OHLC data from a 5 second sqlite DB file.
\enumerate{
 \item \code{fn}         The file name and path of the 5 second sqlite database file
 \item \code{ticker}     The ticker name ("ES")
 \item \code{startSecs}  Start point of data to load
 \item \code{endSecs}    End point of data to load
 \item \code{numMinutes} Bar size to return
}
For \code{KalmanFilter()} Perform simple Kalman Filter
\enumerate{
 \item \code{S}       A single column xts object to operate on, e.g. Cl(OHLC)
 \item \code{gain}    Desired gain.
}
For \code{Kama()}  Kaufman Adaptive Moving Average
\enumerate{
 \item \code{s}       A single column xts object to operate on, e.g. Cl(OHLC)
 \item \code{fast}    The fast move average length
 \item \code{slow}    The slow moving average length
 \item \code{period}  The lookback period
 \item \code{debug}   boolean flag for debug print outs
}
}
\description{
Extensions to TTR by Stergios Marinopoulos
}
\note{
\code{PMO} calculates the DecisionPoint Price Momentum Oscillator.  It is
derived by taking a one day rate of change and smoothing it with two
Exponential Moving Averages. The first EMA defaults to 35 periods, and it is
then further smoothed by a second EMA which defaults to 20 periods.

A note on how the \code{ROC()} indicator from the \code{TTR} package is used
by PMO.  \code{ROC()} provides the percentage difference of a series over
two observations, while the momentum indicator simply provides the
difference.  It offers two different type of computations:
\enumerate{
  \item type="continuous"  \code{ROC=diff(log(price), 1)}          (default)
  \item type="discrete"    \code{ROC=(price/lag.xts(price)) - 1}
}
The EXCEL code for the 1 DAY ROC (Column B) found in the example spreadsheat
uses the "discrete" type and is as follows:\cr\cr
\code{=SUM(((B13/B12)*100)-100)}\cr


\code{MesaStochastic} calculates a John F. Eller's HighPass, Filter, and
Stochastics.  For the \code{MesaStochastic()} function, the reference can be
found is the TASC January 2014 Magazine article 'Predictive Indicators for
effective trading strategies' - The \code{C#} source code for the article is
found in \code{http://www.ninjatrader.com/SC/January2014SC.zip}


\code{SuperSmoother} calculates a John F. Eller's \code{SuperSmoother()} function.
For the \code{SuperSMoother()} function, the reference can be
found is the TASC January 2014 Magazine article 'Predictive Indicators for
effective trading strategies' - The \code{C#} source code for the article is
found in \code{http://www.ninjatrader.com/SC/January2014SC.zip}
}
\references{
#'\enumerate{
 \item DecisionPoint PMO is documented at \url{http://stockcharts.com/school/doku.php?st=pmo&id=chart_school:technical_indicators:dppmo}
 \item TASC January 2014 Magazine article 'Predictive Indicators for Effective Trading Strategies' - The Ehler's \code{C#} source code for the article is found in \url{http://www.ninjatrader.com/SC/January2014SC.zip}
}
}

